// src/types/Ebike.ts
export interface Ebike {
  id: number;
  manufacturer: string;
  model: string;
  zero_60: number;
  topspeed: number;
  curb_weight: number;
  horsepower: number;
  torque: number;
  price_start: number;
  picture: string;
  wiki: string;
  website: string;
  promo_video: string;
  summary: string;
}